+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- Changed `Stage` class and function component types to be the same ([#225])
+
 
 ## [1.0.0-beta.10] - 2021-05-30
 
@@ -455,6 +458,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [0.2.0]: https://github.com/michalochman/react-pixi-fiber/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/michalochman/react-pixi-fiber/compare/v0.1.0...v0.1.1
 
+[#225]: https://github.com/michalochman/react-pixi-fiber/issues/225
 [#221]: https://github.com/michalochman/react-pixi-fiber/issues/221
 [#210]: https://github.com/michalochman/react-pixi-fiber/issues/210
 [#204]: https://github.com/michalochman/react-pixi-fiber/issues/204